All views added to associated with them folders; also work with response refactored
