+++ src/main/java/com/colorit/backend/common/AbstractResponseMaker.java
@@ -3,8 +3,8 @@ package com.colorit.backend.common;
 import com.colorit.backend.entities.UserEntity;
 import com.colorit.backend.services.responses.UserServiceResponse;
 import com.colorit.backend.services.statuses.UserServiceStatus;
-import com.colorit.backend.views.ResponseView;
-import com.colorit.backend.views.UserView;
+import com.colorit.backend.views.output.ResponseView;
+import com.colorit.backend.views.output.UserView;
 import com.colorit.backend.views.ViewStatus;
 import com.colorit.backend.views.ViewStatusCode;
 import org.springframework.context.MessageSource;
@@ -12,6 +12,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
+import javax.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
@@ -25,17 +26,15 @@ import java.util.Locale;
 @Component
 public class AbstractResponseMaker {
 
-    public ResponseEntity<ResponseView> makeResponse(UserServiceResponse userServiceResponse, MessageSource messageSource, Locale locale) {
-        return null;
+    private final @NotNull MessageSource messageSource;
+
+    public AbstractResponseMaker(@NotNull MessageSource messageSource) {
+        this.messageSource = messageSource;
     }
 
-    public ResponseEntity<ResponseView> makeResponse(ViewStatus viewStatus, MessageSource messageSource, Locale locale) {
-        final ResponseView responseView = new ResponseView();
-        for (ViewStatusCode code : viewStatus.getErrors()) {
-            responseView.addError(code.getField(),
-                    messageSource.getMessage(code.getMessage(), null, locale));
-        }
-        return new ResponseEntity<>(responseView, HttpStatus.BAD_REQUEST);
+    // methods package private
+    MessageSource getMessageSource() {
+        return this.messageSource;
     }
 
     ResponseEntity<ResponseView> makeResponse(UserServiceResponse userServiceResponse, String message) {
@@ -70,5 +69,25 @@ public class AbstractResponseMaker {
             return new ResponseEntity<>(responseView, httpStatus);
         }
         return new ResponseEntity<>(new ResponseView<>(), httpStatus);
+    }
+
+    public ResponseEntity<ResponseView> makeUnauthorizedResponse(Locale locale) {
+        final ResponseView responseView = new ResponseView();
+        responseView.addError("general", messageSource.getMessage("unauthorized", null, locale));
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(responseView);
+    }
+
+    // public methods
+    public ResponseEntity<ResponseView> makeResponse(UserServiceResponse userServiceResponse, Locale locale) {
+        return null;
+    }
+
+    public ResponseEntity<ResponseView> makeResponse(ViewStatus viewStatus, Locale locale) {
+        final ResponseView responseView = new ResponseView();
+        for (ViewStatusCode code : viewStatus.getErrors()) {
+            responseView.addError(code.getField(),
+                    messageSource.getMessage(code.getMessage(), null, locale));
+        }
+        return new ResponseEntity<>(responseView, HttpStatus.BAD_REQUEST);
     }
 }
