+++ src/main/java/com/colorit/backend/common/UserResponseMaker.java
@@ -1,19 +1,23 @@
 package com.colorit.backend.common;
 
 import com.colorit.backend.services.responses.UserServiceResponse;
-import com.colorit.backend.views.ResponseView;
+import com.colorit.backend.views.output.ResponseView;
 import org.springframework.context.MessageSource;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
+import javax.validation.constraints.NotNull;
 import java.util.Locale;
 
 @Component
 public class UserResponseMaker extends AbstractResponseMaker {
+    public UserResponseMaker(@NotNull MessageSource messageSource) {
+        super(messageSource);
+    }
+
     @Override
-    public ResponseEntity<ResponseView> makeResponse(UserServiceResponse userServiceResponse,
-                                                     MessageSource messageSource, Locale locale) {
-        final String message = messageSource.getMessage(
+    public ResponseEntity<ResponseView> makeResponse(UserServiceResponse userServiceResponse, Locale locale) {
+        final String message = getMessageSource().getMessage(
                 userServiceResponse.getStatus().getAlternativeMessage(), null, locale);
         return makeResponse(userServiceResponse, message);
     }
