+++ src/main/java/com/colorit/backend/views/input/SignUpView.java
@@ -1,20 +1,23 @@
-package com.colorit.backend.views;
+package com.colorit.backend.views.input;
 
 import com.colorit.backend.entities.UserEntity;
+import com.colorit.backend.views.ViewStatus;
+import com.colorit.backend.views.ViewStatusCode;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class UpdateView extends AbstractView {
+public class SignUpView extends AbstractView {
     private final String nickname;
     private final String email;
     private final String password;
     private final String passwordCheck;
 
+    @SuppressWarnings("unused")
     @JsonCreator
-    public UpdateView(@JsonProperty("nickname") String nickname,
-                      @JsonProperty("email") String email,
-                      @JsonProperty("password") String password,
-                      @JsonProperty("passwordCheck") String passwordCheck) {
+    SignUpView(@JsonProperty("nickname") String nickname,
+               @JsonProperty("email") String email,
+               @JsonProperty("password") String password,
+               @JsonProperty("passwordCheck") String passwordCheck) {
         this.nickname = nickname;
         this.email = email;
         this.password = password;
@@ -24,10 +27,11 @@ public class UpdateView extends AbstractView {
     @Override
     public ViewStatus checkValid() {
         final ViewStatus viewStatus = new ViewStatus();
-
-        if (email != null) {
-            checkEmail(viewStatus, email);
+        if (nickname == null || nickname.isEmpty()) {
+            viewStatus.addStatusCode(ViewStatusCode.EMPTY_NICKNAME);
         }
+
+        checkEmail(viewStatus, email);
 
         boolean passwordFilled = true;
         if (password == null || password.isEmpty()) {
@@ -43,6 +47,7 @@ public class UpdateView extends AbstractView {
         if (passwordFilled && !passwordCheck.equals(password)) {
             viewStatus.addStatusCode(ViewStatusCode.PASSWORD_NOT_MATCH_STATE);
         }
+
         return viewStatus;
     }
