+++ src/main/java/com/colorit/backend/views/input/UpdateView.java
@@ -1,21 +1,22 @@
-package com.colorit.backend.views;
+package com.colorit.backend.views.input;
 
 import com.colorit.backend.entities.UserEntity;
+import com.colorit.backend.views.ViewStatus;
+import com.colorit.backend.views.ViewStatusCode;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class SignUpView extends AbstractView {
+public class UpdateView extends AbstractView {
     private final String nickname;
     private final String email;
     private final String password;
     private final String passwordCheck;
 
-    @SuppressWarnings("unused")
     @JsonCreator
-    SignUpView(@JsonProperty("nickname") String nickname,
-               @JsonProperty("email") String email,
-               @JsonProperty("password") String password,
-               @JsonProperty("passwordCheck") String passwordCheck) {
+    public UpdateView(@JsonProperty("nickname") String nickname,
+                      @JsonProperty("email") String email,
+                      @JsonProperty("password") String password,
+                      @JsonProperty("passwordCheck") String passwordCheck) {
         this.nickname = nickname;
         this.email = email;
         this.password = password;
@@ -25,11 +26,10 @@ public class SignUpView extends AbstractView {
     @Override
     public ViewStatus checkValid() {
         final ViewStatus viewStatus = new ViewStatus();
-        if (nickname == null || nickname.isEmpty()) {
-            viewStatus.addStatusCode(ViewStatusCode.EMPTY_NICKNAME);
-        }
 
-        checkEmail(viewStatus, email);
+        if (email != null) {
+            checkEmail(viewStatus, email);
+        }
 
         boolean passwordFilled = true;
         if (password == null || password.isEmpty()) {
@@ -45,7 +45,6 @@ public class SignUpView extends AbstractView {
         if (passwordFilled && !passwordCheck.equals(password)) {
             viewStatus.addStatusCode(ViewStatusCode.PASSWORD_NOT_MATCH_STATE);
         }
-
         return viewStatus;
     }
