+++ src/main/java/com/colorit/backend/views/output/ResponseView.java
@@ -1,4 +1,4 @@
-package com.colorit.backend.views;
+package com.colorit.backend.views.output;
 
 import java.util.HashMap;
 import java.util.Map;
